Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,6 +1,6 @@
 <stdio.h>
 int main (){
-    int a , b ,c , Avg;
+    float a , b ,c , Avg;
     printf("");
     scanf("%d",&a);
     printf("");
